LPS-162430 Rename ObjectDetails to EditObjectDetails to follow a pattern
