+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ObjectDetails/ScopeContainer.d.ts
@@ -15,7 +15,7 @@
 /// <reference types="react" />
 
 import {FormError} from '@liferay/object-js-components-web';
-import {KeyValuePair} from './ObjectDetails';
+import {KeyValuePair} from './EditObjectDetails';
 import './ObjectDetails.scss';
 interface ScopeContainerProps {
 	companyKeyValuePair: KeyValuePair[];
